requirements.txt
asgiref==3.2.5
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
awsebcli==3.17.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
blessed==1.17.2
botocore==1.14.17
bs4==0.0.1
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
cryptography==2.8
Django==3.0.4
djangorestframework==3.11.0
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
isort==4.3.21
jmespath==0.9.5
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
pathspec==0.5.9
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
wrapt==1.11.2
